Crime scene report #596475
"As long as she was on the scene he could not take any action without a horrible exposure of the
woman whom he loved. But the instant that she was gone he realised how crushing a misfortune this
would be for you, and how all-important it was to set it right. He rushed down, just as he was, in
his bare feet, opened the window, sprang out into the snow, and ran down the lane, where he could
see a dark figure in the moonlight. Sir George Burnwell tried to get away, but Arthur caught him,
and there was a struggle between them, your lad tugging at one side of the coronet, and his opponent
at the other. In the scuffle, your son struck Sir George and cut him over the eye. Then something
suddenly snapped, and your son, finding that he had the coronet in his hands, rushed back, closed
the window, ascended to your room, and had just observed that the coronet had been twisted in the
struggle and was endeavouring to straighten it when you appeared upon the scene."

"Is it possible?" gasped the banker.