Crime scene report #866249
"And now I have a very strange experience to tell you. I had, as you know, cut off my hair in
London, and I had placed it in a great coil at the bottom of my trunk. One evening, after the child
was in bed, I began to amuse myself by examining the furniture of my room and by rearranging my own
little things. There was an old chest of drawers in the room, the two upper ones empty and open, the
lower one locked. I had filled the first two with my linen, and as I had still much to pack away I
was naturally annoyed at not having the use of the third drawer. It struck me that it might have
been fastened by a mere oversight, so I took out my bunch of keys and tried to open it. The very
first key fitted to perfection, and I drew the drawer open. There was only one thing in it, but I am
sure that you would never guess what it was. It was my coil of hair.

"I took it up and examined it. It was of the same peculiar tint, and the same thickness. But then
the impossibility of the thing obtruded itself upon me. How could my hair have been locked in the
drawer? With trembling hands I undid my trunk, turned out the contents, and drew from the bottom my
own hair. I laid the two tresses together, and I assure you that they were identical. Was it not
extraordinary? Puzzle as I would, I could make nothing at all of what it meant. I returned the
strange hair to the drawer, and I said nothing of the matter to the Rucastles as I felt that I had
put myself in the wrong by opening a drawer which they had locked.