Crime scene report #392309
"'I quite follow you,' said I. 'The only point which I could not quite understand was what use you
could make of a hydraulic press in excavating fuller's-earth, which, as I understand, is dug out
like gravel from a pit.'

"'Ah!' said he carelessly, 'we have our own process. We compress the earth into bricks, so as to
remove them without revealing what they are. But that is a mere detail. I have taken you fully into
my confidence now, Mr. Hatherley, and I have shown you how I trust you.' He rose as he spoke. 'I
shall expect you, then, at Eyford at 11.15.'

"'I shall certainly be there.'

"'And not a word to a soul.' He looked at me with a last long, questioning gaze, and then, pressing
my hand in a cold, dank grasp, he hurried from the room.

"Well, when I came to think it all over in cool blood I was very much astonished, as you may both
think, at this sudden commission which had been intrusted to me. On the one hand, of course, I was
glad, for the fee was at least tenfold what I should have asked had I set a price upon my own
services, and it was possible that this order might lead to other ones. On the other hand, the face
and manner of my patron had made an unpleasant impression upon me, and I could not think that his
explanation of the fuller's-earth was sufficient to explain the necessity for my coming at midnight,
and his extreme anxiety lest I should tell anyone of my errand. However, I threw all fears to the
winds, ate a hearty supper, drove to Paddington, and started off, having obeyed to the letter the
injunction as to holding my tongue.

"At Reading I had to change not only my carriage but my station. However, I was in time for the last
train to Eyford, and I reached the little dim-lit station after eleven o'clock. I was the only
passenger who got out there, and there was no one upon the platform save a single sleepy porter with
a lantern. As I passed out through the wicket gate, however, I found my acquaintance of the morning
waiting in the shadow upon the other side. Without a word he grasped my arm and hurried me into a
carriage, the door of which was standing open. He drew up the windows on either side, tapped on the
wood-work, and away we went as fast as the horse could go."