Crime scene report #751356
"Very likely not. It is most important, all the same. As to the note, it is important also, or at
least the initials are, so I congratulate you again."

"I've wasted time enough," said Lestrade, rising. "I believe in hard work and not in sitting by the
fire spinning fine theories. Good-day, Mr. Holmes, and we shall see which gets to the bottom of the
matter first." He gathered up the garments, thrust them into the bag, and made for the door.

"Just one hint to you, Lestrade," drawled Holmes before his rival vanished; "I will tell you the
true solution of the matter. Lady St. Simon is a myth. There is not, and there never has been, any
such person."