Crime scene report #775195
We made our way downstairs as quietly as possible, and out into the bright morning sunshine. In the
road stood our horse and trap, with the half-clad stable-boy waiting at the head. We both sprang in,
and away we dashed down the London Road. A few country carts were stirring, bearing in vegetables to
the metropolis, but the lines of villas on either side were as silent and lifeless as some city in a
dream.

"It has been in some points a singular case," said Holmes, flicking the horse on into a gallop. "I
confess that I have been as blind as a mole, but it is better to learn wisdom late than never to
learn it at all."

In town the earliest risers were just beginning to look sleepily from their windows as we drove
through the streets of the Surrey side. Passing down the Waterloo Bridge Road we crossed over the
river, and dashing up Wellington Street wheeled sharply to the right and found ourselves in Bow
Street. Sherlock Holmes was well known to the force, and the two constables at the door saluted him.
One of them held the horse's head while the other led us in.