Crime scene report #430541
"But his wife--you said that she had ceased to love him."

"This hat has not been brushed for weeks. When I see you, my dear Watson, with a week's accumulation
of dust upon your hat, and when your wife allows you to go out in such a state, I shall fear that
you also have been unfortunate enough to lose your wife's affection."

"But he might be a bachelor."

"Nay, he was bringing home the goose as a peace-offering to his wife. Remember the card upon the
bird's leg."

"You have an answer to everything. But how on earth do you deduce that the gas is not laid on in his
house?"

"One tallow stain, or even two, might come by chance; but when I see no less than five, I think that
there can be little doubt that the individual must be brought into frequent contact with burning
tallow--walks upstairs at night probably with his hat in one hand and a guttering candle in the
other. Anyhow, he never got tallow-stains from a gas-jet. Are you satisfied?"