Crime scene report #930513
"Oh, yes, you can. This is my point." He placed his finger in the centre of the circle. "This is
where we shall find them."

"But the twelve-mile drive?" gasped Hatherley.

"Six out and six back. Nothing simpler. You say yourself that the horse was fresh and glossy when
you got in. How could it be that if it had gone twelve miles over heavy roads?"

"Indeed, it is a likely ruse enough," observed Bradstreet thoughtfully. "Of course there can be no
doubt as to the nature of this gang."

"None at all," said Holmes. "They are coiners on a large scale, and have used the machine to form
the amalgam which has taken the place of silver."