Crime scene report #775456
"I assure you, Watson, without affectation, that the status of my client is a matter of less moment
to me than the interest of his case. It is just possible, however, that that also may not be wanting
in this new investigation. You have been reading the papers diligently of late, have you not?"

"It looks like it," said I ruefully, pointing to a huge bundle in the corner. "I have had nothing
else to do."

"It is fortunate, for you will perhaps be able to post me up. I read nothing except the criminal
news and the agony column. The latter is always instructive. But if you have followed recent events
so closely you must have read about Lord St. Simon and his wedding?"

"Oh, yes, with the deepest interest."

"That is well. The letter which I hold in my hand is from Lord St. Simon. I will read it to you, and
in return you must turn over these papers and let me have whatever bears upon the matter. This is
what he says: