Crime scene report #194629
"Eight weeks passed away like this, and I had written about Abbots and Archery and Armour and
Architecture and Attica, and hoped with diligence that I might get on to the B's before very long.
It cost me something in foolscap, and I had pretty nearly filled a shelf with my writings. And then
suddenly the whole business came to an end."

"To an end?"

"Yes, sir. And no later than this morning. I went to my work as usual at ten o'clock, but the door
was shut and locked, with a little square of cardboard hammered on to the middle of the panel with a
tack. Here it is, and you can read for yourself."

He held up a piece of white cardboard about the size of a sheet of note-paper. It read in this
fashion:

The Red-headed League          is       Dissolved   October 9, 1890.

Sherlock Holmes and I surveyed this curt announcement and the rueful face behind it, until the
comical side of the affair so completely overtopped every other consideration that we both burst out
into a roar of laughter.