Crime scene report #169270
"Quite an interesting study, that maiden," he observed. "I found her more interesting than her
little problem, which, by the way, is rather a trite one. You will find parallel cases, if you
consult my index, in Andover in '77, and there was something of the sort at The Hague last year. Old
as is the idea, however, there were one or two details which were new to me. But the maiden herself
was most instructive."

"You appeared to read a good deal upon her which was quite invisible to me," I remarked.