Crime scene report #996463
"Well, you can imagine how hard it was to settle down to arduous work at £2 a week when I knew that
I could earn as much in a day by smearing my face with a little paint, laying my cap on the ground,
and sitting still. It was a long fight between my pride and the money, but the dollars won at last,
and I threw up reporting and sat day after day in the corner which I had first chosen, inspiring
pity by my ghastly face and filling my pockets with coppers. Only one man knew my secret. He was the
keeper of a low den in which I used to lodge in Swandam Lane, where I could every morning emerge as
a squalid beggar and in the evenings transform myself into a well-dressed man about town. This
fellow, a Lascar, was well paid by me for his rooms, so that I knew that my secret was safe in his
possession.

"Well, very soon I found that I was saving considerable sums of money. I do not mean that any beggar
in the streets of London could earn £700 a year--which is less than my average takings--but I had
exceptional advantages in my power of making up, and also in a facility of repartee, which improved
by practice and made me quite a recognised character in the City. All day a stream of pennies,
varied by silver, poured in upon me, and it was a very bad day in which I failed to take £2.

"As I grew richer I grew more ambitious, took a house in the country, and eventually married,
without anyone having a suspicion as to my real occupation. My dear wife knew that I had business in
the City. She little knew what.