Crime scene report #866097
"No, but I observe the second half of a return ticket in the palm of your left glove. You must have
started early, and yet you had a good drive in a dog-cart, along heavy roads, before you reached the
station."

The lady gave a violent start and stared in bewilderment at my companion.

"There is no mystery, my dear madam," said he, smiling. "The left arm of your jacket is spattered
with mud in no less than seven places. The marks are perfectly fresh. There is no vehicle save a
dog-cart which throws up mud in that way, and then only when you sit on the left-hand side of the
driver."