Crime scene report #623616
"My sister had told me some weeks before that I might have the pick of her geese for a Christmas
present, and I knew that she was always as good as her word. I would take my goose now, and in it I
would carry my stone to Kilburn. There was a little shed in the yard, and behind this I drove one of
the birds--a fine big one, white, with a barred tail. I caught it, and prying its bill open, I
thrust the stone down its throat as far as my finger could reach. The bird gave a gulp, and I felt
the stone pass along its gullet and down into its crop. But the creature flapped and struggled, and
out came my sister to know what was the matter. As I turned to speak to her the brute broke loose
and fluttered off among the others.

"'Whatever were you doing with that bird, Jem?' says she.